Use BigInteger for stake amounts (#325)
Also fixes CBW-1611.
Now the minimum stake amount is not checked
if not changing the amount while updating the status.

---------

+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/baker/BakerRegistrationConfirmationActivity.kt
@@ -17,7 +17,6 @@ import com.concordium.wallet.ui.account.accountdetails.AccountDetailsActivity
 import com.concordium.wallet.ui.bakerdelegation.common.BaseDelegationBakerActivity
 import com.concordium.wallet.util.UnitConvertUtil
 import com.concordium.wallet.util.dropAfterDecimalPlaces
-import com.concordium.wallet.util.toBigInteger
 import java.math.BigInteger
 
 class BakerRegistrationConfirmationActivity : BaseDelegationBakerActivity() {
@@ -77,7 +76,7 @@ class BakerRegistrationConfirmationActivity : BaseDelegationBakerActivity() {
 
             UPDATE_BAKER_KEYS -> {
                 viewModel.bakerDelegationData.amount =
-                    viewModel.bakerDelegationData.account?.accountBaker?.stakedAmount.toBigInteger()
+                    viewModel.bakerDelegationData.account?.accountBaker?.stakedAmount
                 updateViewsUpdateBakerKeys()
             }
